MOL-326: change container_id to be a str instead of UUID.
The container_id doesn't need to be a UUID and it will be a str when accessing overføringspakke to be downloaded to on-prem.

Changes are also made to reflect that the logic of container_id should be contained in the domain layer and not in the sas_generator client.

This entails changes of UUID to str in some other functions also included in the commit